requirements.txt
openai>=2.16.0
pandas>=3.0.0
faiss-cpu>=1.13.2
numpy>=2.4.2
